move get_sorted_plot_dicts to configuration parser
